chore(clippy): clear all warnings to unblock CI
CI's Clippy job had been failing for 5 consecutive commits because
clippy 1.91 promoted several lints from allow → warn and the
project's pipeline runs `-D warnings`. Failing CI was blocking the
Build & Publish Image deploy that our env-separation fix needs to
take effect. None of these are behavior changes; all are mechanical
lint fixes.

- src/api/background_watcher.rs: drop redundant `.trim()` before
  `.split_whitespace()` (split_whitespace handles whitespace already);
  extract the three `format!("{base}.…")` paths from inside the outer
  `format!` so the `format_in_format_args` lint stops complaining;
  inline the `$pid` shell var reference instead of `{pid_val}`.
- src/api/mission_runner.rs: collapse the manual `for idx in
  0..=haystack.len() - needle.len()` find into an iterator
  `.find(…)`.
- src/k8s_pod.rs: drop the redundant `(exit_code as i32)` cast —
  `output.status.code()` already returns `i32`.
- src/library/mod.rs: allow `only_used_in_recursion` on
  `collect_skill_files_recursive`'s `&self` — the helper is a method
  for namespacing only.
- src/api/system.rs: hoist the `DEPLOY_DEBOUNCE_SECS` constant sanity
  check from a runtime `assert!(true)` to a `const _: () = assert!(…)`
  so it evaluates at compile time and the optimizer can't drop it.
- src/bin/bg_watchd.rs: tighten the doc-list continuation indent from
  4 spaces to 2 spaces (`doc_overindented_list_items`).

cargo fmt + cargo clippy --workspace --all-targets clean.

+++ src/k8s_pod.rs
@@ -663,7 +663,7 @@ impl K8sPodClient {
             "k8s_pod exec_command done"
         );
 
-        let status = ExitStatus::from_raw((exit_code as i32) << 8);
+        let status = ExitStatus::from_raw(exit_code << 8);
         Ok(Output {
             status,
             stdout: output.stdout,
